move duplicated logic to utils
